cli_apps/pip_data/newtags/requirements.txt
about-time
agate
agate-dbf
agate-excel
agate-sql
aiohttp
aiopg
aiosignal
alabaster
alembic
alive-progress
amqp
ansiwrap
anyio
apipkg
appdirs
APScheduler
apsw
argcomplete
args
arrow
asgiref
asteval
asttokens
async-generator
async-timeout
attrs
auto-tagify
autocommand
Automat
Babel
backcall
backports.csv
beautifulsoup4
billiard
binaryornot
black
bleach
blessed
blinker
blis
bottle
Brotli
bs4
BTrees
btrfsutil
build
CacheControl
cached-property
cachelib
cachetools
cairocffi
case
catalogue
cement
certifi
cffi
Chameleon
chardet
charset-normalizer
cheap-repr
cli-helpers
click
click-default-group-wheel
click-didyoumean
click-plugins
click-repl
clint
colander
colorama
colorlog
Colr
columnize
commonmark
configobj
constantly
contextlib2
courlan
coverage
croniter
cryptography
cssselect
csvkit
cymem
Cython
dateparser
dbfread
debtcollector
decompyle3
decorator
deform
defusedxml
Deprecated
distlib
distro
Django
django3-all-access
dmenu
dmenu-extended
dnspython
docopt
docutils
Elixir
email-validator
english-words
erl
et-xmlfile
evdev
eventlet
exceptiongroup
executing
fancycompleter
fastjsonschema
feedfinder2
feedparser
filelock
fire
flake8
Flask
Flask-Humanize
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
fluidity-sm
frozenlist
future
gevent
goose-extractor
goose3
gprof2dot
grab
grapheme
graphviz
greenlet
gunicorn
h11
h2
hiredis
howdoi
hpack
html2text
html5-parser
html5lib
htmldate
httpcore
httpx
huey
humanize
hupper
hyperframe
hyperlink
icecream
idna
imagesize
importlib-metadata
incremental
inflect
iniconfig
invoke
iotop
ipython
iso8601
isodate
isort
itemadapter
itemloaders
itsdangerous
jaraco.context
jaraco.functools
jaraco.text
jedi
jeepney
jellyfish
jieba
jieba3k
Jinja2
jinja2-time
jmespath
joblib
jrnl
json5
jusText
keep
keyboard
keyring
keystoneauth1
langcodes
langdetect
leather
lexicon
libfdt
libvirt-python
line-profiler
litecli
littleutils
lockfile
Logria
loguru
lxml
Mako
Markdown
markdown-it-py
markdownify
MarkupSafe
matplotlib-inline
mccabe
mdurl
mdv
MechanicalSoup
mechanize
more-itertools
MouseInfo
mrq
msgpack
multi-rake
multidict
murmurhash
mutagen
mycli
mypy-extensions
mysql-connector
mysql-connector-python
mysqlclient
nanoid
natsort
netaddr
netifaces
networkx
newspaper3k
nftables
nltk
nose
nspektr
numpy
oauthlib
objgraph
olefile
openpyxl
ordered-set
os-service-types
oslo.config
oslo.i18n
oslo.serialization
oslo.utils
outcome
outdated
packaging
pandas
pangocairocffi
pangocffi
parsedatetime
parsel
parso
passwordgenerator
PasteDeploy
pathspec
pathy
pbr
pdbr
peewee
pep517
peppercorn
persistent
pexpect
pickleshare
pika
Pillow
pipx
pixcat
pkginfo
plan
plaster
plaster-pastedeploy
platformdirs
pluggy
ply
powerline-status
poyo
preshed
print-schema
priority
procrastinate
progress
prometheus-client
prompt-toolkit
Protego
protobuf
psutil
psycopg2-binary
ptyprocess
pulsemixer
pure-eval
pwquality
py
pyaes
pyasn1
pyasn1-modules
pycairo
pycld2
pycodestyle
pycparser
pycrypto
pycryptodomex
pycurl
pydantic
PyDispatcher
pydub
pyficache
pyfiglet
pyflakes
PyGetWindow
PyGithub
Pygments
PyGObject
PyJWT
pyKey
PyMsgBox
PyMySQL
PyNaCl
pynput
pyparsing
pyperclip
pyproject_hooks
PyQt5
PyQt5-Qt5
PyQt5-sip
pyquery
pyramid
pyramid-debugtoolbar
pyramid-jinja2
pyramid-mako
pyramid-retry
pyramid-tm
pyramid-zodbconn
PyRect
pyrepl
pyrsistent
PyScreeze
PySocks
pyspider
pytermgui
pytest
pytest-cov
pythemis
python-crontab
python-daemon
python-dateutil
python-dotenv
python-keystoneclient
python-Levenshtein
python-oembed
python-slugify
python-swiftclient
python-xlib
python3-xlib
pytils
pytimeparse
pytube
pytube3
pytweening
pytz
pytz-deprecation-shim
pyxdg
PyYAML
questionary
queuelib
ranger-fm
readability-lxml
readme-renderer
redbird
redis
regex
repoze.lru
requests
requests-file
requests-futures
requests-oauthlib
requests-toolbelt
resolvelib
retrying
rfc3986
rich
rich-rst
rocketry
rq-dashboard
schedule
schema
scipy
scrape-search-engine
segtok
selection
selenium
service-identity
sgmllib3k
shellingham
six
smart-open
smdv
sniffio
snoop
snowballstemmer
sortedcontainers
soupsieve
spacy-legacy
spacy-loggers
spark-parser
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
SQLAlchemy
sqlcipher3-binary
sqlite-fts4
sqlite-utils
sqlparse
srsly
stack-data
stem
stevedore
sultan
summa
systemd-python
tabulate
tblib
team
tenacity
term-background
termcolor
terminaltables
terminator
text-unidecode
textblob
textual
textwrap3
thefuzz
thinc
timeloop
tinydb
tinysegmenter
tld
tldextract
tomber
toml
tomli
tornado
torrequest
tqdm
tracer
traitlets
transaction
translationstring
trash-cli
trepan3k
trio
trio-websocket
trove-classifiers
TUIFIManager
twine
Twisted
typer
typing_extensions
tzdata
tzlocal
u-msgpack-python
uc-micro-py
ujson
uncompyle6
Uni-Curses
urllib3
urwid
urwid-readline
user-agent
userpath
validate
validate-pyproject
validators
venusian
vine
virtualenv
w3lib
waitress
wasabi
watchdog
wcwidth
webdriver-manager
webencodings
weblib
WebOb
websockets
WebTest
Werkzeug
wmctrl
wrapt
WsgiDAV
wsproto
WTForms
x256
xdis
xlrd
yake
yarl
youtube-dl
youtube-search
youtube-search-python
zc.lockfile
ZConfig
zdaemon
ZEO
zipp
ZODB
zodbpickle
zodburi
zope.deprecation
zope.event
zope.interface
